backend/requirements.txt
asgiref==3.8.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Django==5.1
django-environ==0.11.2
django-rest-framework==0.1.0
django-storages==1.14.4
djangorestframework==3.15.2
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
h11==0.14.0
idna==3.7
jwt==1.3.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
psycopg==3.2.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
